Add files via upload
chgd event to delete net from any mouse button press, to right button press. REASON: when working with scene zoomed way out, and dragging node(s),, the user might accidently press on a net and unknowingly delete it, This happened becuz the left button press was used for bth drag and for delete. So I chg the event that deleted net to right button. Now dragging is safe.24jan2022-12:30